ConDec-359: Rename StrategyProvider class into PersistenceProvider and add a method that returns the PersistenceManager for a given decision knowledge element
[decision] The getPersistenceStrategy method is still included for "first class" decision knowledge elements that are independently documented. Thus, we still have the Strategy design pattern for autarkical decision knowledge. [/decision]
