+++ src/main/java/de/uhd/ifi/se/decision/management/jira/model/DecisionKnowledgeProject.java
@@ -5,7 +5,7 @@ import java.util.Set;
 import de.uhd.ifi.se.decision.management.jira.persistence.AbstractPersistenceManager;
 import de.uhd.ifi.se.decision.management.jira.persistence.ActiveObjectPersistence;
 import de.uhd.ifi.se.decision.management.jira.persistence.JiraIssuePersistence;
-import de.uhd.ifi.se.decision.management.jira.persistence.StrategyProvider;
+import de.uhd.ifi.se.decision.management.jira.persistence.PersistenceProvider;
 
 /**
  * Interface for a project and its configuration. The project is a JIRA project
@@ -117,7 +117,7 @@ public interface DecisionKnowledgeProject {
 	 *         strategy is the default strategy.
 	 */
 	static AbstractPersistenceManager getPersistenceStrategy(String projectKey) {
-		return StrategyProvider.getPersistenceStrategy(projectKey);
+		return PersistenceProvider.getPersistenceStrategy(projectKey);
 	}
 
 	/**
