+++ src/test/java/de/uhd/ifi/se/decision/management/jira/view/treeviewer/TestTreeViewer.java
@@ -30,7 +30,7 @@ import de.uhd.ifi.se.decision.management.jira.model.DecisionKnowledgeElement;
 import de.uhd.ifi.se.decision.management.jira.model.DecisionKnowledgeElementImpl;
 import de.uhd.ifi.se.decision.management.jira.model.KnowledgeType;
 import de.uhd.ifi.se.decision.management.jira.persistence.AbstractPersistenceManager;
-import de.uhd.ifi.se.decision.management.jira.persistence.StrategyProvider;
+import de.uhd.ifi.se.decision.management.jira.persistence.PersistenceProvider;
 import net.java.ao.EntityManager;
 import net.java.ao.test.jdbc.NonTransactional;
 import net.java.ao.test.junit.ActiveObjectsJUnitRunner;
@@ -64,7 +64,7 @@ public class TestTreeViewer extends TestSetUpWithIssues {
 		treeViewer.setCheckCallback(checkCallback);
 		treeViewer.setThemes(themes);
 		treeViewer.setData(data);
-		persistenceStrategy = StrategyProvider.getPersistenceStrategy("TEST");
+		persistenceStrategy = PersistenceProvider.getPersistenceStrategy("TEST");
 	}
 
 	@Test
